# user-microservice/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: User Microservice
  description: |-
    This is the User Microservice for the EasyConf scenario in CSE2115. It is responsible for managing users, events (including their tracks and attendees), and invitations.
  version: 0.0.0
servers:
  - url: http://localhost:8085
    description: Production server for integration.
...
paths:
...
  /track:
    post:
      tags:
        - Track Management
      summary: Create a new track.
      description: This operation can only be performed by the general chair of the related event.
      operationId: addTrack
      requestBody:
        description: Track object to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Track'
      responses:
        '200':
          description: successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '401':
          description: Unauthorized access.
        '400':
          description: Invalid input for new track
      security:
        - api_key: [ ]

    put:
      tags:
        - Track Management
      summary: Update a track.
      description: This operation can only be performed by the track's PC Chair, or the general chair of the associated event.
      operationId: updateTrack
      requestBody:
        description: Track data to be changed.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Track'
      responses:
        '204':
          description: Track successfully updated.
        '400':
          description: Invalid Track object supplied.
        '401':
          description: Unauthorized access.
        '404':
          description: Track object not found.
      security:
        - api_key: [ ]
    get:
      tags:
        - Track Management
      summary: Gets tracks
      operationId: getTrack
      parameters:
        - name: name
          in: query
          description: Track's name.
          required: false
          explode: true
          schema:
            type: string
        - name: eventID
          in: query
          description: Event to which the track belongs.
          required: false
          explode: true
          schema:
            type: integer
        - name: paper_type
          in: query
          description: Track's paper type
          required: false
          explode: true
          schema:
            $ref: '#/components/schemas/PaperType'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
        '400':
          description: Invalid query values.
        '401':
          description: Unauthorized access.
      security:
        - api_key: [ ]
  /track/{trackID}:
    get:
      tags:
        - Track Management
      summary: Get track by ID
      description: This operation can only be performed by the general chair of the related event.
      operationId: getTrackByID
      parameters:
        - name: trackID
          in: path
          description: ID of the track to be returned.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Invalid TrackID supplied
        '401':
          description: Unahtorized access.
        '404':
          description: Track not found
      security:
        - api_key: [ ]

    delete:
      tags:
        - Track Management
      summary: Delete a track.
      operationId: deleteTrack
      parameters:
        - name: trackID
          in: path
          description: ID of the track to be deleted.
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Track successfully deleted.
        '400':
          description: Invalid trackID supplied.
        '401':
          description: Unahtorized access.
        '404':
          description: Track not found.
      security:
        - api_key: [ ]
...
components:
  schemas:
...
    Track:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        title:
          type: string
          example: "track Title 1"
        description:
          type: string
          example: "track 1 is ..., aims to ..."
        submit_deadline:
          type: string
          format: date
          description: the deadline for all paper submissions
          example: "2022/11/24, 21:59"
        review_deadline:
          type: string
          format: date
          description: the deadline for all paper reviews
          example: "2022/11/24, 21:59"
        paper_type:
          $ref: '#/components/schemas/PaperType'
        event_id:
          type: integer
          format: int64
          example: 5
...
    PaperType:
      type: string
      enum:
        - full-paper
        - short-paper
        - position-paper
...
  securitySchemes:
    api_key:
      type: http
      scheme: bearer
      bearerFormat: JWT